+++ packages/dumb/src/models/print/pdf-data.model.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024, Diabeloop
+ * Copyright (c) 2023-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -30,7 +30,9 @@ import {
   type DeviceConfig,
   type ParameterConfig,
   type PumpConfig,
-  type BasicData
+  type BasicData,
+  MobileAppConfig,
+  SecurityBasalConfig
 } from 'medical-domain'
 
 export interface PdfData {
@@ -47,7 +49,9 @@ export interface PdfSettingsData extends PdfData {
     device?: DeviceConfig
     pump?: PumpConfig
     cgm?: CgmConfig
+    mobileApplication?: MobileAppConfig
     parameters?: ParameterConfig[]
+    securityBasals?: SecurityBasalConfig
   }
   originalDate: string
 }