+++ packages/dumb/src/modules/print/settings-print-view/settings-print-view.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -29,21 +29,31 @@ import { type PdfDocumentOverridden } from '../../../models/print/pdf-override.m
 import { type PdfSettingsData } from '../../../models/print/pdf-data.model'
 import { type PrintViewParams } from '../../../models/print/print-view-params.model'
 import { PrintView } from '../print-view/print-view'
-import { type CgmConfig, type DeviceConfig, type PumpConfig, type TimePrefs } from 'medical-domain'
+import {
+  type CgmConfig,
+  type DeviceConfig,
+  DeviceSystem,
+  MobileAppConfig,
+  type PumpConfig,
+  type TimePrefs
+} from 'medical-domain'
 import { type DeviceMetadata } from '../../../models/device-metadata.model'
 import i18next from 'i18next'
 import {
   getDeviceMetadata,
   getDeviceParametersTableData,
   getParametersByLevel,
+  getSafetyBasalProfileTableData,
   getTableDataByDataType
 } from './settings-print-view.util'
 import {
   type BasicSettingsTable,
   type ParameterSettingsTableRow,
+  SafetyBasalProfileTableRow,
   type SettingsTableColumn
 } from '../../../models/print/pdf-settings-table.model'
 import { PdfSettingsDataType } from '../../../models/enums/pdf-settings-data-type.enum'
+import { getSafetyBasalItems } from '../../../utils/safety-basal-profile/safety-basal-profile.util'
 
 const t = i18next.t.bind(i18next)
 
@@ -69,14 +79,23 @@ export class SettingsPrintView extends PrintView<PdfSettingsData> {
   }
 
   render(): void {
-    this.renderTableSection(PdfSettingsDataType.Device)
+    if (this.data.payload?.device?.name === DeviceSystem.Dblg1) {
+      this.renderTableSection(PdfSettingsDataType.Device)
+    }
+    if (this.data.payload?.mobileApplication?.identifier.toUpperCase() === DeviceSystem.Dblg2) {
+      this.renderTableSection(PdfSettingsDataType.MobileApplication)
+    }
     this.renderTableSection(PdfSettingsDataType.Pump)
     this.renderTableSection(PdfSettingsDataType.Cgm)
     this.renderDeviceParametersTableSection()
+    this.renderSafetyBasalProfileTableSection()
     this.resetText()
   }
 
-  private renderSettingsSection(tableData: BasicSettingsTable, width: number, optionalParams?: { zebra: boolean, showHeaders: boolean }): void {
+  private renderSettingsSection(tableData: BasicSettingsTable, width: number, optionalParams?: {
+    zebra: boolean,
+    showHeaders: boolean
+  }): void {
     this.renderTableHeading(tableData.heading, {
       columnsDefaults: {
         width,
@@ -95,7 +114,7 @@ export class SettingsPrintView extends PrintView<PdfSettingsData> {
     }, tableData.columns, tableData.rows)
   }
 
-  private readonly getDataByDataType = (type: PdfSettingsDataType): CgmConfig | DeviceConfig | PumpConfig | undefined => {
+  private readonly getDataByDataType = (type: PdfSettingsDataType): CgmConfig | DeviceConfig | PumpConfig | MobileAppConfig | undefined => {
     switch (type) {
       case PdfSettingsDataType.Cgm:
         return this.data.payload?.cgm
@@ -103,6 +122,8 @@ export class SettingsPrintView extends PrintView<PdfSettingsData> {
         return this.data.payload?.device
       case PdfSettingsDataType.Pump:
         return this.data.payload?.pump
+      case PdfSettingsDataType.MobileApplication:
+        return this.data.payload?.mobileApplication
     }
   }
 
@@ -147,5 +168,21 @@ export class SettingsPrintView extends PrintView<PdfSettingsData> {
 
       this.renderSettingsSection(tableData, this.chartArea.width, { zebra: true, showHeaders: true })
     })
+  }
+
+  private renderSafetyBasalProfileTableSection(): void {
+    const safetyBasalProfile = this.data.payload?.securityBasals ?? null
+    const isSafetyBasalAvailable = !!safetyBasalProfile?.rates?.length && safetyBasalProfile?.rates?.length > 0
+
+    if (!isSafetyBasalAvailable) {
+      return
+    }
+
+    const originalDate = this.data.originalDate ? this.data.normalTime : undefined
+
+    const safetyBasalItems = getSafetyBasalItems(safetyBasalProfile)
+    const tableData = getSafetyBasalProfileTableData(safetyBasalItems as SafetyBasalProfileTableRow[], this.chartArea.width, this.data.timezone, originalDate)
+
+    this.renderSettingsSection(tableData, this.chartArea.width, { zebra: true, showHeaders: true })
   }
 }