+++ packages/dumb/src/modules/print/settings-print-view/settings-print-view.util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023-2024, Diabeloop
+ * Copyright (c) 2023-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -34,15 +34,18 @@ import {
 import {
   type CgmConfig,
   type DeviceConfig,
+  DeviceSystem,
+  MobileAppConfig,
   type ParameterConfig,
   type PumpConfig,
+  SecurityBasalConfig,
   type TimePrefs
 } from 'medical-domain'
 import i18next from 'i18next'
 import {
   type AlignType,
   type ParameterSettingsTable,
-  type ParameterSettingsTableRow,
+  type ParameterSettingsTableRow, SafetyBasalProfileTableRow,
   type SettingsTable,
   type SettingsTableColumn,
   type SettingsTableRow
@@ -73,7 +76,7 @@ const TABLE_COLUMNS: SettingsTableColumn[] = [
   }
 ]
 
-type TableData = CgmConfig | DeviceConfig | PumpConfig
+type TableData = CgmConfig | DeviceConfig | PumpConfig | MobileAppConfig | SecurityBasalConfig
 
 const getTimePrefs = (timezone = TIMEZONE_UTC): TimePrefs => {
   const timezoneName = timezone === TIMEZONE_UTC ? new Intl.DateTimeFormat().resolvedOptions().timeZone : timezone
@@ -104,16 +107,27 @@ const getTextByDataTableType = (type: PdfSettingsDataType): string => {
       return t('Device')
     case PdfSettingsDataType.Pump:
       return t('Pump')
+    case PdfSettingsDataType.MobileApplication:
+      return t('mobile-application')
   }
 }
 
 const getSubTextByDataTableType = (type: PdfSettingsDataType, data: TableData): string | undefined => {
   switch (type) {
     case PdfSettingsDataType.Cgm:
+    case PdfSettingsDataType.MobileApplication:
       return
-    case PdfSettingsDataType.Device:
-    case PdfSettingsDataType.Pump:
-      return `- ${data.name}`
+    case PdfSettingsDataType.Device: {
+      const deviceData = data as DeviceConfig
+      if (deviceData.name === DeviceSystem.Dblg2) {
+        return
+      }
+      return `- ${deviceData.name}`
+    }
+    case PdfSettingsDataType.Pump: {
+      const pumpData = data as PumpConfig
+      return `- ${pumpData.name}`
+    }
   }
 }
 
@@ -170,6 +184,19 @@ const getTableRowsByDataTableType = (type: PdfSettingsDataType, data: TableData,
         label: t('Pump version'),
         value: pump.swVersion
       }]
+    case PdfSettingsDataType.MobileApplication: {
+      const mobileApp = data as MobileAppConfig
+      return [{
+        label: t('Manufacturer'),
+        value: mobileApp.manufacturer
+      }, {
+        label: t('Name'),
+        value: mobileApp.identifier
+      },{
+        label: t('Software version'),
+        value: mobileApp.swVersion
+      }]
+    }
   }
 }
 
@@ -233,6 +260,35 @@ export const getDeviceParametersTableData = (parameters: ParameterSettingsTableR
   }]
 
   return { heading, columns, rows: parameters }
+}
+
+export const getSafetyBasalProfileTableData = (rates: SafetyBasalProfileTableRow[], width: number, timezone?: string, date?: string) => {
+  const timePrefs = getTimePrefs(timezone)
+  const text = t('safety-basal-profile')
+
+  const heading = getTableHeading(text, timePrefs, formatCurrentDate(), date)
+
+  const columns = [{
+    id: 'startTime',
+    header: t('start-time'),
+    cache: false,
+    align: 'left' as AlignType,
+    width: (width * 0.3)
+  }, {
+    id: 'endTime',
+    header: t('end-time'),
+    cache: false,
+    align: 'left' as AlignType,
+    width: (width * 0.3)
+  }, {
+    id: 'rate',
+    header: t('basal-rate'),
+    cache: false,
+    align: 'right' as AlignType,
+    width: (width * 0.4)
+  }]
+
+  return { heading, columns, rows: rates }
 }
 
 export const getParametersByLevel = (parameters?: ParameterConfig[]): Map<number, ParameterSettingsTableRow[]> => {