+++ packages/dumb/src/utils/safety-basal-profile/safety-basal-profile.util.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2024, Diabeloop
+ * Copyright (c) 2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,42 +25,48 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export interface AppConfig {
-  BANNER_ENABLED: boolean
-  BANNER_LABEL_EN: string
-  BANNER_LABEL_ES: string
-  BANNER_LABEL_DE: string
-  BANNER_LABEL_FR: string
-  BANNER_LABEL_IT: string
-  BANNER_LABEL_NL: string
-  VERSION: string
-  API_HOST: string
-  BRANDING: string
-  DOMAIN_NAME?: string
-  DEV: boolean
-  TEST: boolean
-  LATEST_TERMS?: string
-  LATEST_TRAINING?: string
-  LATEST_RELEASE?: string
-  SUPPORT_WEB_ADDRESS?: string
-  CONTACT_SUPPORT_WEB_URL: string
-  ASSETS_URL?: string | null
-  METRICS_SERVICE?: string | null
-  TERMS_PRIVACY_DATE?: string
-  STONLY_WID: string
-  COOKIE_BANNER_CLIENT_ID: string
-  SESSION_TIMEOUT: number
-  YLP820_BASAL_TIME: number
-  CBG_BUCKETS_ENABLED: boolean
-  ECPS_ENABLED: boolean
-  AUTH0_DOMAIN: string
-  AUTH0_ISSUER: string
-  AUTH0_CLIENT_ID: string
-  YLPZ_RA_LAD_FR: string
-  YLPZ_RA_LAD_EN: string
-  YLPZ_RA_LAD_NL: string
-  YLPZ_RA_LAD_IT: string
-  YLPZ_RA_LAD_ES: string
-  YLPZ_RA_LAD_DE: string
-  IDLE_TIMEOUT_MS: number
-}
+import { SecurityBasalConfig } from 'medical-domain'
+import { getSafetyBasalItems } from './safety-basal-profile.util'
+
+describe('SafetyBasalProfileUtil', () => {
+  describe('getSafetyBasalItems', () => {
+    it('should compute end times correctly', () => {
+      const securityBasalConfig: SecurityBasalConfig = {
+        rates: [
+
+          {
+            rate: 1.6,
+            start: 510
+          },
+          {
+            rate: 1,
+            start: 0
+          },
+          {
+            rate: 0.4,
+            start: 840
+          }
+        ]
+      }
+
+      const result = getSafetyBasalItems(securityBasalConfig)
+      expect(result).toEqual([
+        {
+          rate: '1 U/h',
+          startTime: '12:00 AM',
+          endTime: '8:30 AM'
+        },
+        {
+          rate: '1.6 U/h',
+          startTime: '8:30 AM',
+          endTime: '2:00 PM'
+        },
+        {
+          rate: '0.4 U/h',
+          startTime: '2:00 PM',
+          endTime: '12:00 AM'
+        }
+      ])
+    })
+  })
+})