+++ CHANGELOG.md
@@ -2,16 +2,35 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## 3.5.0-beta - 2024-09-03
+## 3.6.0-beta - 2024-09-03
 ### Added
 - YLP-2408 Display basal safety profile in Devices view
+- YLP-2692 Display basal safety profile in the PDF report
+- YLP-2821 As a DBLG2 user, I want to be able to see the device tab
 - YLP-3128 As an HCP, I want to see which device is used by each patient
+- YLP-3145 As a yourloops user I want to do a csv export
 
 ### Changed
 - YLP-3262 Allow redirection to DBLG2 app from YourLoops
+- YLP-3332 Rename "Pump settings" page of PDF export to "Settings"
 
 ### Fixed
+- YLP-2543 PDF report legend label font color is wrong
+- YLP-3390 Meal bolus calculation not taking into account second part of biphasic bolus
+
+### Engineering use
+- YLP-3242 Cache-busting for static files in blip
+
+## 3.5.0 - 2024-11-06
+### Changed
+- YLP-3310 Display Physical Activity below other events in Daily view
+
+### Fixed
+- YLP-3300 Legend in "rescue carbs & manual bolus intakes" table is not consistent with what is displayed
 - YLP-3303 Rescue carbs tooltip in "Rescue carbs & manual bolus intakes" table is wrong
+
+### Engineering use
+- YLP-3306 embed product labeling document (pdf) in yourloops
 
 ## 3.4.1 - 2024-09-25
 ### Fixed